camerar bs

+++ Assets/Scripts/Balls/BallDetectionZone.cs
@@ -1,9 +1,13 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using NaughtyAttributes;
 
 public class BallDetectionZone : MonoBehaviour
 {
+    public bool camera = false;
+    [ShowIf("camera")]
+    public CameraController.CameraState state;
     // Start is called before the first frame update
     void Start()
     {
@@ -17,7 +21,14 @@ public class BallDetectionZone : MonoBehaviour
     }
     public void OnTriggerEnter2D(Collider2D coll) {
         if (coll.gameObject.layer != 3) return;
-        Debug.Log("kill ball");
-        CourseController.Instance.BallDodged(GlobalFunctions.FindComponent<BallController>(coll.gameObject));
+        if (!camera) {
+            Debug.Log("kill ball");
+            CourseController.Instance.BallDodged(GlobalFunctions.FindComponent<BallController>(coll.gameObject));
+        }
+        else {
+            if (state == CameraController.CameraState.Ball && CameraController.Instance.currentCameraState == CameraController.CameraState.Player) return;
+            Debug.Log("tracked "+ state.ToString());
+            CameraController.Instance.SetCameraState(state, GlobalFunctions.FindComponent<BallController>(coll.gameObject));
+        }
     }
 }
